requirements.txt
# Core
yt-dlp
moviepy
python-dotenv
rich

# OpenAI mode (paid)
openai

# Local / Free mode (Colab)
faster-whisper
groq

# Web Frontend
gradio
